+++ src/dotnetCampus.Ipc.Abstractions/IIpcRequestContext.cs
@@ -1,5 +1,4 @@
 ﻿using dotnetCampus.Ipc.Abstractions.Context;
-using dotnetCampus.Ipc.PipeCore.Context;
 
 namespace dotnetCampus.Ipc.Abstractions
 {
