+++ src/dotnetCampus.Ipc.Abstractions/IIpcRequestHandler.cs
