+++ src/dotnetCampus.Ipc.Abstractions/IPeerProxy.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
 using dotnetCampus.Ipc.Abstractions.Context;
-using dotnetCampus.Ipc.PipeCore.Context;
-using dotnetCampus.Ipc.PipeCore.Utils;
 
 namespace dotnetCampus.Ipc.Abstractions
 {
