+++ src/dotnetCampus.Ipc.PipeCore/Context/IpcConfiguration.cs
@@ -1,5 +1,4 @@
 ﻿using dotnetCampus.Ipc.Abstractions;
-using dotnetCampus.Ipc.Abstractions.Context;
 using dotnetCampus.Ipc.PipeCore.Utils;
 
 namespace dotnetCampus.Ipc.PipeCore.Context
@@ -34,25 +33,5 @@ namespace dotnetCampus.Ipc.PipeCore.Context
          */
         public byte[] MessageHeader { set; get; } =
             {0x64, 0x6F, 0x74, 0x6E, 0x65, 0x74, 0x20, 0x63, 0x61, 0x6D, 0x70, 0x75, 0x73};
-    }
-
-    class EmptyIpcRequestHandler : IIpcRequestHandler
-    {
-        public IIpcHandleRequestMessageResult HandleRequestMessage(IIpcRequestContext requestContext)
-        {
-            // 我又不知道业务，不知道怎么玩……
-            var responseMessage = new IpcRequestMessage(nameof(EmptyIpcRequestHandler), new IpcBufferMessage(new byte[0]));
-            return new IpcHandleRequestMessageResult(responseMessage);
-        }
-    }
-
-    class IpcHandleRequestMessageResult : IIpcHandleRequestMessageResult
-    {
-        public IpcHandleRequestMessageResult(IpcRequestMessage returnMessage)
-        {
-            ReturnMessage = returnMessage;
-        }
-
-        public IpcRequestMessage ReturnMessage { get; }
     }
 }
