+++ src/dotnetCampus.Ipc.PipeCore/Context/IpcHandleRequestMessageResult.cs
@@ -1,0 +1,14 @@
+﻿using dotnetCampus.Ipc.Abstractions;
+
+namespace dotnetCampus.Ipc.PipeCore.Context
+{
+    class IpcHandleRequestMessageResult : IIpcHandleRequestMessageResult
+    {
+        public IpcHandleRequestMessageResult(IpcRequestMessage returnMessage)
+        {
+            ReturnMessage = returnMessage;
+        }
+
+        public IpcRequestMessage ReturnMessage { get; }
+    }
+}
