+++ src/dotnetCampus.Ipc.PipeCore/Context/Logger_/ILogger.cs
