+++ src/dotnetCampus.Ipc.PipeCore/Context/Logger_/IpcLogger.cs
@@ -1,0 +1,17 @@
+﻿namespace dotnetCampus.Ipc.PipeCore.Context
+{
+    class IpcLogger : ILogger
+    {
+        public IpcLogger(IpcContext ipcContext)
+        {
+            IpcContext = ipcContext;
+        }
+
+        public override string ToString()
+        {
+            return $"[{IpcContext.PipeName}]";
+        }
+
+        private IpcContext IpcContext { get; }
+    }
+}
