+++ src/dotnetCampus.Ipc.PipeCore/Core_/DebugContext.cs
@@ -1,0 +1,9 @@
+﻿/*
+namespace dotnetCampus.Ipc.PipeCore
+{
+    class DebugContext
+    {
+        public const string DoNotUseAck = "当前不再需要回复ACK信息，因为管道通讯形式是不需要获取对方是否收到";
+    }
+}
+*/
