+++ src/dotnetCampus.Ipc.PipeCore/Core_/IpcMessageConverter.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using dotnetCampus.Ipc.Abstractions.Context;
 using dotnetCampus.Ipc.PipeCore.Context;
 using dotnetCampus.Ipc.PipeCore.Utils;
+using dotnetCampus.Ipc.PipeCore.Utils.Extensions;
 
 namespace dotnetCampus.Ipc.PipeCore
 {
