+++ src/dotnetCampus.Ipc.PipeCore/Core_/IpcPipeServerMessageProvider.cs
@@ -2,11 +2,12 @@
 using System.IO.Pipes;
 using System.Threading.Tasks;
 using dotnetCampus.Ipc.PipeCore.Context;
-using dotnetCampus.Ipc.PipeCore.Utils;
 
 namespace dotnetCampus.Ipc.PipeCore
 {
-    // 提供一个客户端连接
+    /// <summary>
+    /// 提供一个客户端连接
+    /// </summary>
     internal class IpcPipeServerMessageProvider : IDisposable
     {
         public IpcPipeServerMessageProvider(IpcContext ipcContext, IpcServerService ipcServerService)
