+++ src/dotnetCampus.Ipc.PipeCore/Core_/ServerStreamMessageReader.cs
@@ -376,9 +376,4 @@ namespace dotnetCampus.Ipc.PipeCore
             PeerConnectBroke?.Invoke(this, e);
         }
     }
-
-    class DebugContext
-    {
-        public const string DoNotUseAck = "当前不再需要回复ACK信息，因为管道通讯形式是不需要获取对方是否收到";
-    }
 }
