+++ src/dotnetCampus.Ipc/IpcProxy.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
+using dotnetCampus.Ipc.Abstractions;
 using dotnetCampus.Ipc.Context;
 using dotnetCampus.Ipc.Utils;
